Update getting-started-kubernetes-windows.md
I don't think it's appropriate to say that Flannel support is in beta, when it's not merged into flannel master yet. "Beta" might make people (such as me) believe that it's fairly straight-forward to get flannel on Windows working, which it most definetely is not. The correct wording for the time being is imho to state that flannel is in development